Cambios en los nombres de los paquetes que referenciaban erroneamente a wikipedia.
