+++ app/src/main/java/ayds/winchester/songinfo/moredetails/model/MoreDetailsModelInjector.kt
@@ -3,11 +3,11 @@ package ayds.winchester.songinfo.moredetails.model
 import android.content.Context
 import ayds.winchester.songinfo.moredetails.model.repository.CardsRepository
 import ayds.winchester.songinfo.moredetails.model.repository.CardsRepositoryImpl
-import ayds.winchester.songinfo.moredetails.model.repository.local.wikipedia.MoreDetailsDataBase
-import ayds.winchester.songinfo.moredetails.model.repository.local.wikipedia.database.CursorToCardMapperImpl
-import ayds.winchester.songinfo.moredetails.model.repository.local.wikipedia.database.MoreDetailsDataBaseImpl
+import ayds.winchester.songinfo.moredetails.model.repository.local.MoreDetailsDataBase
+import ayds.winchester.songinfo.moredetails.model.repository.local.database.CursorToCardMapperImpl
+import ayds.winchester.songinfo.moredetails.model.repository.local.database.MoreDetailsDataBaseImpl
 import ayds.winchester.songinfo.moredetails.view.MoreDetailsView
-import ayds.winchester.songinfo.moredetails.model.repository.external.wikipedia.BrokerInjector.broker
+import ayds.winchester.songinfo.moredetails.model.repository.external.BrokerInjector.broker
 
 object MoreDetailsModelInjector {
     private lateinit var moreDetailsModel: MoreDetailsModel
