+++ app/src/main/java/ayds/winchester/songinfo/moredetails/model/repository/external/proxys/Proxy.kt
@@ -1,4 +1,4 @@
-package ayds.winchester.songinfo.moredetails.model.repository.external.wikipedia.proxys
+package ayds.winchester.songinfo.moredetails.model.repository.external.proxys
 
 import ayds.winchester.songinfo.moredetails.model.entities.Card
