+++ app/src/main/java/ayds/winchester/songinfo/moredetails/model/repository/external/proxys/ProxyWikipedia.kt
@@ -1,4 +1,4 @@
-package ayds.winchester.songinfo.moredetails.model.repository.external.wikipedia.proxys
+package ayds.winchester.songinfo.moredetails.model.repository.external.proxys
 
 import ayds.winchester.songinfo.moredetails.model.entities.Card
 import ayds.winchester.songinfo.moredetails.model.entities.CardSource
