+++ app/src/test/java/ayds/winchester/songinfo/moredetails/model/repository/external/proxys/ProxyWikipediaTest.kt
@@ -1,7 +1,6 @@
 package ayds.winchester.songinfo.moredetails.model.repository.external.proxys
 
 import ayds.winchester.songinfo.moredetails.model.entities.Card
-import ayds.winchester.songinfo.moredetails.model.repository.external.wikipedia.proxys.ProxyWikipedia
 import ayds.winchester1.wikipedia.WikipediaArtistInfo
 import ayds.winchester1.wikipedia.WikipediaService
 import io.mockk.every
@@ -18,18 +17,7 @@ class ProxyWikipediaTest {
 
     @Test
     fun `on found artist name should return artist info card `() {
-//        val wikipediaArtistInfo: WikipediaArtistInfo = mockk()
-//        val card: Card = mockk()
-//
-//        every { wikipediaService.getArtistInfo("artistName") } returns wikipediaArtistInfo
-//
-//        val result = proxyWikipedia.getCard("artistName")
-//
-//        card
-//
-//
-//
-//        Assert.assertEquals(card, wikipediaCard)
+
     }
 
 }
